add ( detail  ) : add custom painter
